requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pandas
pydantic
python-dotenv
httpx
pytest
celery
redis
